[user] Split User Controller to authorized and unauthorized controllers
